Rename DisposableEventListener to addEventListener since it's no longer an object, but a function
